requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
uvicorn==0.27.1
pydantic==2.10.6
python-dotenv==1.0.1
langchain==0.3.16
langchain-nvidia-ai-endpoints==0.3.9
langchain-community==0.3.16
numexpr==2.9.0
pydantic==2.10.6
chromadb==0.6.3
pytest==8.1.1
jinja2==3.1.6
pandas==2.2.3
ipython==9.0.2
numpy==1.26.4
